revert azure functions
