--- src/basic_func_impl/class_basic_invoke/readme.md
@@ -1,1 +1,0 @@
-PF flow is going to be copied here during deployment.
